Refactor: 예산 임계값 계산 로직 공통화, 빌드 의존성 최적화 및 미사용 웹 코드 제거(Task57)
1. 예산 임계값 판정 로직 공통화 (B1)

수학적 유틸리티 신설: 예산 소비율에 따른 알림 발송 기준을 통일하기 위해 BudgetThresholdMath 클래스를 추가했습니다. (1% 단위 래더 계산, isCrossed, safeRatio 로직 포함)

발행부 연동: 기존 3개의 알림 발행처(Publisher)가 분산 관리하던 임계값 조건을 DEFAULT_MONTHLY_THRESHOLDS 상수 및 BudgetThresholdMath.isCrossed 메서드를 바라보도록 일원화하여 정책 정합성을 높였습니다.

2. 그레이들(Gradle) 빌드 의존성 슬리밍 (A3)

의존성 다이어트: 메인 컴파일 영역에서 불필요하게 선언되어 있던 spring-boot-starter-validation, jackson-databind:2.18.2, jackson-datatype-jsr310:2.18.2 라이브러리를 제거하여 빌드 경량화를 도모했습니다.

테스트 스코프 조정: 일부 테스트 코드에서 JavaTimeModule을 직접 참조하는 케이스를 처리하기 위해 testImplementation 스코프로만 jackson-datatype-jsr310을 남겨두었으며, 버전은 Spring Boot BOM 관리에 위임했습니다.

3. 도메인 구조 정돈 및 예외 핸들링 이관 (B2)

패키지 정렬: 특정 컨트롤러에 종속적이던 MeResponse 객체를 공용 api.dto 패키지로 이동시켜 재사용성을 높였습니다.

예외 처리 단일화: 비즈니스 레이어에서 발생하던 IllegalArgumentException에 대한 400(Bad Request) 응답 규격 처리 로직을 BillingRestExceptionHandler로 이관하여 전역 예외 처리 메커니즘과 포맷을 일치시켰습니다.

4. 프론트엔드 레거시 데드코드 청소 (A1)

더 이상 사용되지 않는 웹 프론트엔드 영역의 유틸 파일 3종(client-fetch.ts, types.ts, utils.ts)을 영구 삭제했습니다. 프로젝트 전반의 정적 분석 결과 참조 건수가 0건임을 사전에 확인했습니다.

5. 안정성 검증

백엔드 영역의 모든 컴파일(compileJava / compileTestJava) 및 전체 단위 테스트(./gradlew test)를 수행하여 BUILD SUCCESSFUL을 기록했습니다. 외부 비즈니스 로직의 본질, MQ 페이로드 스키마, HTTP 노출 계약은 기존과 완벽히 동일하게 보존됩니다.
