+++ services/billing-service/src/main/java/com/eevee/billingservice/api/dto/MeResponse.java
@@ -1,0 +1,4 @@
+package com.eevee.billingservice.api.dto;
+
+public record MeResponse(String userId) {
+}
