+++ services/billing-service/src/main/java/com/eevee/billingservice/service/BudgetThresholdEventPublisher.java
@@ -12,11 +12,9 @@ import org.springframework.amqp.rabbit.core.RabbitTemplate;
 import org.springframework.stereotype.Component;
 
 import java.math.BigDecimal;
-import java.math.RoundingMode;
 import java.nio.charset.StandardCharsets;
 import java.time.Instant;
 import java.time.LocalDate;
-import java.util.ArrayList;
 import java.util.List;
 
 @Component
@@ -24,25 +22,7 @@ public class BudgetThresholdEventPublisher {
 
     private static final Logger log = LoggerFactory.getLogger(BudgetThresholdEventPublisher.class);
 
-    /**
-     * Monthly spend / budget ratios at which one notification is emitted when first crossed (1% steps).
-     * <p>
-     * NOTE: This is intentionally generated from integer percents to avoid floating-point drift.
-     * When switching back to 10% steps, change {@code PERCENT_STEP} to 10.
-     */
-    private static final int PERCENT_STEP = 1;
-    private static final List<BigDecimal> DEFAULT_THRESHOLDS = buildPercentThresholdsInclusive(PERCENT_STEP, 100, PERCENT_STEP);
-
-    private static List<BigDecimal> buildPercentThresholdsInclusive(int fromPercent, int toPercent, int stepPercent) {
-        if (fromPercent <= 0 || toPercent <= 0 || toPercent < fromPercent || stepPercent <= 0) {
-            throw new IllegalArgumentException("Invalid percent threshold range");
-        }
-        List<BigDecimal> thresholds = new ArrayList<>((toPercent - fromPercent) / stepPercent + 1);
-        for (int pct = fromPercent; pct <= toPercent; pct += stepPercent) {
-            thresholds.add(BigDecimal.valueOf(pct).movePointLeft(2));
-        }
-        return List.copyOf(thresholds);
-    }
+    private static final List<BigDecimal> DEFAULT_THRESHOLDS = BudgetThresholdMath.DEFAULT_MONTHLY_THRESHOLDS;
 
     private final RabbitTemplate rabbitTemplate;
     private final BillingRabbitProperties rabbitProperties;
@@ -83,7 +63,7 @@ public class BudgetThresholdEventPublisher {
         }
 
         for (BigDecimal threshold : DEFAULT_THRESHOLDS) {
-            if (!isCrossed(monthlyTotalUsdBefore, monthlyTotalUsdAfter, monthlyBudgetUsd, threshold)) {
+            if (!BudgetThresholdMath.isCrossed(monthlyTotalUsdBefore, monthlyTotalUsdAfter, monthlyBudgetUsd, threshold)) {
                 continue;
             }
             publishOne(out, userId, teamId, apiKeyId, apiKeyAlias, monthStart, threshold, monthlyTotalUsdAfter, monthlyBudgetUsd);
@@ -140,24 +120,6 @@ public class BudgetThresholdEventPublisher {
         }
     }
 
-    private static boolean isCrossed(
-            BigDecimal beforeTotalUsd,
-            BigDecimal afterTotalUsd,
-            BigDecimal budgetUsd,
-            BigDecimal thresholdPct
-    ) {
-        BigDecimal beforePct = safeRatio(beforeTotalUsd, budgetUsd);
-        BigDecimal afterPct = safeRatio(afterTotalUsd, budgetUsd);
-        return beforePct.compareTo(thresholdPct) < 0 && afterPct.compareTo(thresholdPct) >= 0;
-    }
-
-    private static BigDecimal safeRatio(BigDecimal total, BigDecimal budget) {
-        if (total == null || budget == null || budget.signum() <= 0) {
-            return BigDecimal.ZERO;
-        }
-        return total.divide(budget, 8, RoundingMode.HALF_UP);
-    }
-
     private static String resolveSubjectType(String userId, String teamId, String apiKeyId) {
         if (hasText(apiKeyId)) {
             return "API_KEY";
@@ -172,4 +134,3 @@ public class BudgetThresholdEventPublisher {
         return value != null && !value.isBlank();
     }
 }
-
