--- services/billing-service/web/src/lib/utils.ts
@@ -1,1 +1,0 @@
-export { cn } from "@ai-usage/ui";
